feat(bencoding): keep source around when decoding
Doing this with a `ReaderT` around a `Ref String` right now. If we need
to keep more state we'll switch to a record or something, but I don't
think it's needed right now.

The `id3as/purescript-parsing` repo needs a PR for `MonadReader` &
`MonadAsk` instances to automatically be defined for `ParserT`.